Associate projects and activities
Creates a one-to-many relationship between projects and activities, and
provides all entry points in project controller to add activities to
projects and remove activities from projects.

The activity controller still exposes entries to get a specific activity
and edit it, but since there are no activities without a project, this
controller does not offer entries to add or remove acitivities directly.