Create TimeRecord entity
Correlates TimeRecord entities with KowalskiUser and Task ones.
Each appointment is directly associated to an user and a task. At this
moment the system will not allow creating time records associated to
activities. The time spent in a given activity will be determined by the
sum of all related tasks.